+++ hyperdbg/include/SDK/Imports/User/HyperDbgLibImports.h
@@ -73,10 +73,33 @@ IMPORT_EXPORT_LIBHYPERDBG BOOLEAN
 hyperdbg_u_check_multiline_command(CHAR * current_command, BOOLEAN reset);
 
 IMPORT_EXPORT_LIBHYPERDBG BOOLEAN
-hyperdbg_u_set_custom_driver_path(CHAR * DriverFilePath, CHAR * DriverName);
+hyperdbg_u_set_custom_driver_path(CHAR * driver_file_path, CHAR * driver_name);
 
 IMPORT_EXPORT_LIBHYPERDBG VOID
 hyperdbg_u_use_default_driver_path();
+
+//
+// Reading memory
+//
+IMPORT_EXPORT_LIBHYPERDBG BOOLEAN
+hyperdbg_u_read_memory(UINT64                              target_address,
+                       DEBUGGER_READ_MEMORY_TYPE           memory_type,
+                       DEBUGGER_READ_READING_TYPE          reading_Type,
+                       UINT32                              pid,
+                       UINT32                              size,
+                       BOOLEAN                             get_address_mode,
+                       DEBUGGER_READ_MEMORY_ADDRESS_MODE * address_mode,
+                       BYTE *                              target_buffer_to_store,
+                       UINT32 *                            return_length);
+
+IMPORT_EXPORT_LIBHYPERDBG VOID
+hyperdbg_u_show_memory_or_disassemble(DEBUGGER_SHOW_MEMORY_STYLE   style,
+                                      UINT64                       address,
+                                      DEBUGGER_READ_MEMORY_TYPE    memory_type,
+                                      DEBUGGER_READ_READING_TYPE   reading_type,
+                                      UINT32                       pid,
+                                      UINT32                       size,
+                                      PDEBUGGER_DT_COMMAND_OPTIONS dt_details);
 
 //
 // Connect to local or remote debugger
@@ -107,7 +130,7 @@ hyperdbg_u_pause_debuggee();
 // Exported functionality of the 'bp' command
 //
 VOID
-hyperdbg_u_set_breakpoint(UINT64 Address, UINT32 Pid, UINT32 Tid, UINT32 CoreNumer);
+hyperdbg_u_set_breakpoint(UINT64 address, UINT32 pid, UINT32 tid, UINT32 core_numer);
 
 #ifdef __cplusplus
 }