+++ hyperdbg/libhyperdbg/code/debugger/commands/meta-commands/script.cpp
@@ -221,6 +221,8 @@ HyperDbgScriptReadFileAndExecuteCommand(std::vector<std::string> & PathAndArgs)
 
 /**
  * @brief Parsing the command line options for scripts
+ * @param argc
+ * @param argv
  *
  * @return INT
  */